Change single collection torch data format   torch.float16 -> torch.bfloat16

+++ forge/test/operators/utils/test_data.py
@@ -348,6 +348,6 @@ class TestCollectionTorch:
 
     single = TestCollection(
         dev_data_formats=[
-            torch.float16,
+            torch.bfloat16,
         ],
     )
